Move quantum4bpmn JSON to resource folder in client
