+++ client/src/app/tabs/bpmn/modeler/BpmnModeler.js
@@ -43,7 +43,7 @@ import 'bpmn-js/dist/assets/diagram-js.css';
 import 'bpmn-js/dist/assets/bpmn-font/css/bpmn-embedded.css';
 
 import 'diagram-js-minimap/assets/diagram-js-minimap.css';
-import quantMEModdleExtension from './extension/quantum4bpmn.json';
+import quantMEModdleExtension from '../../../../../resources/quantme/quantum4bpmn.json';
 
 
 export default class CamundaBpmnModeler extends BpmnModeler {
